move LabelTab out of js/diary/list
Since LabelTab is the parent component of everything else in /diary now, it makes sense for this to live at the root of /diary.
