+++ www/js/diary/LabelTab.tsx
@@ -3,20 +3,20 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { View, useWindowDimensions } from "react-native";
 import { FlashList } from "@shopify/flash-list";
-import TripCard from "../cards/TripCard";
-import PlaceCard from "../cards/PlaceCard";
-import UntrackedTimeCard from "../cards/UntrackedTimeCard";
-import { angularize, getAngularService } from "../../angular-react-helper";
-import LoadMoreButton from "./LoadMoreButton";
-import useAppConfig from "../../useAppConfig";
+import TripCard from "./cards/TripCard";
+import PlaceCard from "./cards/PlaceCard";
+import UntrackedTimeCard from "./cards/UntrackedTimeCard";
+import { angularize, getAngularService } from "../angular-react-helper";
+import LoadMoreButton from "./list/LoadMoreButton";
+import useAppConfig from "../useAppConfig";
 import { useTranslation } from "react-i18next";
-import { invalidateMaps } from "../../components/LeafletView";
+import { invalidateMaps } from "../components/LeafletView";
 import { ActivityIndicator, Appbar } from "react-native-paper";
-import FilterSelect from "./FilterSelect";
-import DateSelect from "./DateSelect";
+import FilterSelect from "./list/FilterSelect";
+import DateSelect from "./list/DateSelect";
 import Bottleneck from "bottleneck";
 import moment from "moment";
-import TimelineScrollList from "./TimelineScrollList";
+import TimelineScrollList from "./list/TimelineScrollList";
 
 let labelPopulateFactory, labelsResultMap, notesResultMap, showPlaces;
 const placeLimiter = new Bottleneck({ maxConcurrent: 2, minTime: 500 });
